requirements.txt
gradio
huggingface_hub
tavily-python
